Remove `TFNumber` class and extends `UInt8` from `Number` directly.
